Move Gateway API to api.rs
