+++ sui/src/lib.rs
@@ -2,6 +2,7 @@
 // Copyright (c) 2022, Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod api;
 pub mod benchmark;
 pub mod config;
 pub mod keystore;
